#vfb-135 . Move back skeletons for neuron to THREEDCanvas, needs renderer to use library when fetching it
